Charge and power indication: half way

+++ lib/ui/box/box.h
@@ -6,27 +6,27 @@
 struct Box
 {
     private:
-        const uint16_t topLeftCornerX;
-        const uint16_t topLeftCornerY;
-        const uint8_t width;
-        const uint8_t height;
+        const int16_t topLeftCornerX;
+        const int16_t topLeftCornerY;
+        const uint16_t width;
+        const uint16_t height;
         const uint8_t padding;
         const uint8_t border;
 
     public:
         Box(
-            const uint16_t topLeftCornerX,
-            const uint16_t topLeftCornerY,
-            const uint8_t width,
-            const uint8_t height, 
+            const int16_t topLeftCornerX,
+            const int16_t topLeftCornerY,
+            const uint16_t width,
+            const uint16_t height, 
             const uint8_t padding,
             const uint8_t border
         );
 
-        uint16_t x(); 
-        uint16_t y(); 
-        uint8_t wdth();
-        uint8_t hght();
+        int16_t x(); 
+        int16_t y(); 
+        uint16_t wdth();
+        uint16_t hght();
         uint8_t pad();
         uint8_t brd();
 };
